requirements.txt
Django
psycopg2
django-allauth
djangorestframework
django-rest-auth
djangorestframework-gis
